Add tests and rename my test files
